code files reorganization
