+++ calista/engines/spark.py
@@ -24,7 +24,7 @@ from pyspark.sql.window import Window
 
 import calista.core._conditions as cond
 import calista.core.rules as R
-import calista.label.label_spark as label
+import calista.label.spark.iban.label_spark as label
 from calista.core._aggregate_conditions import Count, Max, Mean, Median, Min, Sum
 from calista.core.aggregates import AggregateDataset
 from calista.core.catalogue import PythonTypes
